backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
distlib==0.3.9
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
oauthlib==3.2.2
openai==1.78.0
platformdirs==4.3.8
psycopg==3.2.6
psycopg-binary==3.2.6
pydantic==2.11.4
pydantic_core==2.33.2
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.3.0
virtualenv==20.31.2
